backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic-settings
python-jose[cryptography]
bcrypt
python-multipart
aiofiles
sse-starlette
pytest
pytest-asyncio
httpx
